// projekat-FRONT/src/app/models/user.model.ts
export class User {
    id?:number;
    username: string;           //ovo postaje email
    password: string;
    firstname: string;
    lastname: string;
    adress: string;
    city: string;
    country: string;
    phoneNumber: string;
    uloga: string;

    //constructor(uloga : string){ this.uloga = uloga}
}